pipeline/natl_state_estimates/requirements.txt
git+https://github.com/COVID-IWG/epimargin@master#egg=adaptive
appnope==0.1.0
astroid==2.4.1
attrs==19.3.0
backcall==0.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.5.0
cycler==0.10.0
decorator==4.4.2
Fiona==1.8.13.post1
geopandas==0.7.0
google-cloud-storage==1.7.0
idna==2.9
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.0
joblib==0.14.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.2.1
mccabe==0.6.1
munch==2.5.0
numpy==1.18.2
pandas==1.0.3
parso==0.7.0
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
pylint==2.5.2
pyparsing==2.4.7
pyproj==2.6.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.0
Shapely==1.7.0
six==1.14.0
sklearn==0.0
statsmodels==0.11.1
tikzplotlib==0.9.1
toml==0.10.1
tqdm==4.45.0
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.9
wcwidth==0.1.9
wrapt==1.12.1
